Add support for Magnetics Inversion (maginv3d_60) GIFtools group
