+++ geoh5py/groups/__init__.py
@@ -26,6 +26,8 @@ from .container import ContainerGroup
 from .custom import CustomGroup
 from .drillhole import DrillholeGroup, IntegratorDrillholeGroup
 from .giftools.giftools import GIFtoolsGroup
+from .giftools.base import BaseGIFtoolsGroup
+from .giftools.magnetics_inversion import MagInv3D
 from .giftools.octree_inversion import DCOctreeInversion
 from .group_type import GroupType
 from .integrator import (
